+++ packages/web-components/core/src/animation/controlsToPromise.ts
@@ -1,0 +1,9 @@
+import type { AnimationPlaybackControlsWithThen } from "motion";
+
+export async function controlsToPromise(
+  controls: AnimationPlaybackControlsWithThen,
+) {
+  return new Promise((resolve) => {
+    controls.then(() => resolve(null));
+  });
+}