+++ packages/web-components/core/src/animation/index.ts
@@ -1,0 +1,2 @@
+export * from "./controlsToPromise.js";
+export * from "./awaitAllAnimations.js";