+++ packages/web-components/core/src/index.ts
@@ -1,1 +1,3 @@
-export { createFocusTrap } from "./focusTrap";
+export * from "./focusTrap/index.js";
+export * from "./promise/index.js";
+export * from "./animation/index.js";