+++ packages/web-components/modal/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brytdesigns/web-component-modal
 
+## 1.0.4
+
+### Patch Changes
+
+- Introduced new animation utilities
+- Updated dependencies
+  - @brytdesigns/web-component-core@1.0.1
+
 ## 1.0.3
 
 ### Patch Changes